Remove utilities from models.base
